Revert `AlphaVersionBanner` changes.
