+++ app/lib/main/sharezone.dart
@@ -22,6 +22,7 @@ import 'package:sharezone/dynamic_links/dynamic_link_bloc.dart';
 import 'package:sharezone/dynamic_links/dynamic_links.dart';
 import 'package:sharezone/main/auth_app.dart';
 import 'package:sharezone/main/bloc_dependencies.dart';
+import 'package:sharezone/main/constants.dart';
 import 'package:sharezone/main/sharezone_app.dart';
 import 'package:sharezone/main/sharezone_bloc_providers.dart';
 import 'package:sharezone/navigation/logic/navigation_bloc.dart';
@@ -29,8 +30,8 @@ import 'package:sharezone/notifications/notifications_permission.dart';
 import 'package:sharezone/onboarding/group_onboarding/logic/signed_up_bloc.dart';
 import 'package:sharezone/sharezone_plus/subscription_service/subscription_flag.dart';
 import 'package:sharezone/util/flavor.dart';
+import 'package:sharezone/widgets/alpha_version_banner.dart';
 import 'package:sharezone/widgets/animation/color_fade_in.dart';
-import 'package:sharezone/widgets/deployment_stage_banner.dart';
 import 'package:sharezone_utils/device_information_manager.dart';
 import 'package:sharezone_widgets/sharezone_widgets.dart';
 
@@ -111,7 +112,8 @@ class _SharezoneState extends State<Sharezone> with WidgetsBindingObserver {
             textDirection: TextDirection.ltr,
             child: _ThemeSettingsProvider(
               blocDependencies: widget.blocDependencies,
-              child: DeploymentStageBanner(
+              child: AlphaVersionBanner(
+                enabled: kDevelopmentStage == 'ALPHA',
                 child: Stack(
                   children: [
                     MultiProvider(
